requirements.txt
Flask==2.3.3
Flask-Login==0.6.2
Flask-SQLAlchemy==3.0.5
Flask-WTF==1.1.1
Werkzeug==2.3.7
numpy==1.24.3
pandas==1.5.3
plotly==5.16.1
reportlab==4.0.4
python-dotenv==1.0.0
openai==0.28.0
gunicorn==21.2.0
psycopg2-binary==2.9.7
email-validator==1.3.1 